Extract common config
