+++ client/configs/webpack.config.prod.js
@@ -1,0 +1,5 @@
+const { config } = require('./common');
+
+module.exports = {
+  ...config,
+};
